chore: relax condition to emit back pc for Pandar64
Previously the point cloud back would only be emitted if ring 0
(the top most ring for Pandar64) would have a point. This implied
that in open scenes where ring 0 has no points within 0.1 and 200m,
no point cloud would be emitted back by the SDK.

Here we change the logic so that we always emit back a point cloud
regardless of points being in ring 0 or not.